+++ src/main/java/roomescape/reservation/controller/ReservationTimeApiController.java
@@ -1,4 +1,4 @@
-package roomescape.time.controller;
+package roomescape.reservation.controller;
 
 import java.net.URI;
 import java.util.List;
@@ -9,9 +9,9 @@ import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RestController;
-import roomescape.time.dto.TimeRequest;
-import roomescape.time.dto.TimeResponse;
-import roomescape.time.service.ReservationTimeService;
+import roomescape.reservation.dto.TimeRequest;
+import roomescape.reservation.dto.TimeResponse;
+import roomescape.reservation.service.ReservationTimeService;
 
 @RestController
 public class ReservationTimeApiController {
