+++ src/main/java/roomescape/reservation/dto/ReservationRequest.java
@@ -3,8 +3,8 @@ package roomescape.reservation.dto;
 import java.time.LocalDate;
 import roomescape.reservation.domain.Name;
 import roomescape.reservation.domain.Reservation;
-import roomescape.theme.domain.Theme;
-import roomescape.time.domain.ReservationTime;
+import roomescape.reservation.domain.ReservationTime;
+import roomescape.reservation.domain.Theme;
 
 public record ReservationRequest(String name, LocalDate date, Long themeId, Long timeId) {
