+++ src/main/java/roomescape/reservation/dto/ReservationResponse.java
@@ -2,8 +2,6 @@ package roomescape.reservation.dto;
 
 import java.time.LocalDate;
 import roomescape.reservation.domain.Reservation;
-import roomescape.theme.dto.ThemeResponse;
-import roomescape.time.dto.TimeResponse;
 
 public record ReservationResponse(Long id, String name, LocalDate date, ThemeResponse theme, TimeResponse time) {
