+++ src/test/java/roomescape/reservation/controller/ReservationTimePageControllerTest.java
@@ -1,4 +1,4 @@
-package roomescape.theme.controller;
+package roomescape.reservation.controller;
 
 import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.get;
 import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.status;
@@ -8,19 +8,24 @@ import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.autoconfigure.web.servlet.WebMvcTest;
+import org.springframework.boot.test.mock.mockito.MockBean;
 import org.springframework.test.web.servlet.MockMvc;
+import roomescape.reservation.service.ReservationService;
 
-@WebMvcTest(ThemePageController.class)
-class ThemePageControllerTest {
+@WebMvcTest(ReservationTimePageController.class)
+class ReservationTimePageControllerTest {
+
+    @MockBean
+    private ReservationService reservationService;
 
     @Autowired
     private MockMvc mockMvc;
 
     @Test
-    @DisplayName("/theme 을 요청하면 admin/theme.html 를 반환한다.")
-    void requestTheme() throws Exception {
-        mockMvc.perform(get("/admin/theme"))
+    @DisplayName("/time 을 요청하면 time.html 를 반환한다.")
+    void requestTime() throws Exception {
+        mockMvc.perform(get("/admin/time"))
                 .andExpect(status().isOk())
-                .andExpect(view().name("admin/theme"));
+                .andExpect(view().name("admin/time"));
     }
 }
